Moved view functions from presenter to view helpers
