--- presenters/presenters_suite_test.go
@@ -1,13 +1,0 @@
-package presenters_test
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
-)
-
-func TestPresenters(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Presenters Suite")
-}
